requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pillow
google-genai==0.3.0
bleach==6.1.0
jsonschema==4.20.0
python-telegram-bot==20.7
python-dotenv==1.0.0